Sprint3/requirements.txt
Flask
Flask-MySQLdb
Flask-Login
boto3
python-dotenv
mysqlclient
SQLAlchemy
PyMySQL
openpyxl
fpdf2
psutil
PyPDF2
